refactor(init): extract iOS parsing into IosProjectParser
`init.dart` mixed orchestration with low-level pbxproj and Info.plist
parsing. Extract the parsing into a dedicated `IosProjectParser` in
`utils/`, mirroring `GradleParser`. Both methods are pure (string in,
value out), so they're directly unit-testable. `init.dart` keeps the
file-finding and detection orchestration but is no longer responsible
for pbxproj graph traversal or Info.plist regex.

Net: -107 lines from init.dart, +118 lines in the new utility.
TESTING_NOTES.md flagged this exact extraction as a prerequisite for
unit tests; the next change in this series adds those tests.
